+++ packages/ice/src/commands/start.ts
@@ -1,17 +1,17 @@
 import * as path from 'path';
+import detectPort from 'detect-port';
 import type { Configuration as DevServerConfiguration } from 'webpack-dev-server';
 import type { Context, TaskConfig } from 'build-scripts';
 import type { StatsError, Compiler, Configuration } from 'webpack';
 import type { Config } from '@ice/webpack-config/esm/types';
 import type { AppConfig, RenderMode } from '@ice/runtime';
 import type ora from '@ice/bundles/compiled/ora/index.js';
-
 import WebpackDevServer from '@ice/bundles/compiled/webpack-dev-server/lib/Server.js';
 import webpack from '@ice/bundles/compiled/webpack/index.js';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
 import { getWebpackConfig } from '@ice/webpack-config';
 import type { ExtendsPluginAPI, ServerCompiler, GetAppConfig, GetRoutesConfig, GetDataloaderConfig } from '../types';
-import { IMPORT_META_RENDERER, IMPORT_META_TARGET, RUNTIME_TMP_DIR, WEB } from '../constant.js';
+import { IMPORT_META_RENDERER, IMPORT_META_TARGET, RUNTIME_TMP_DIR, WEB, DEFAULT_HOST, DEFAULT_PORT } from '../constant.js';
 import webpackCompiler from '../service/webpackCompiler.js';
 import formatWebpackMessages from '../utils/formatWebpackMessages.js';
 import prepareURLs from '../utils/prepareURLs.js';
@@ -123,15 +123,25 @@ async function startDevServer({
   routeManifest,
 }: StartDevServerOptions): Promise<{ compiler: Compiler; devServer: WebpackDevServer }> {
   const { commandArgs, userConfig, rootDir, applyHook, extendsPluginAPI: { serverCompileTask } } = context;
-  const { port, host, https = false } = commandArgs;
   const { ssg, ssr } = userConfig;
   const { getAppConfig } = hooksAPI;
   const webTaskConfig = taskConfigs.find(({ name }) => name === WEB);
   const customMiddlewares = webpackConfigs[0].devServer?.setupMiddlewares;
+  // Get the value of the host and port from the command line, environment variables, and webpack config.
+  // Value priority: process.env.PORT > commandArgs > webpackConfig > DEFAULT.
+  const host = process.env.HOST ||
+    commandArgs.host ||
+    webpackConfigs[0].devServer?.host ||
+    DEFAULT_HOST;
+  const port = process.env.PORT ||
+    commandArgs.port ||
+    webpackConfigs[0].devServer?.port ||
+    await detectPort(DEFAULT_PORT);
+
+
   let devServerConfig: DevServerConfiguration = {
     port,
     host,
-    https,
     setupMiddlewares: (middlewares, devServer) => {
       let renderMode: RenderMode;
       // If ssr is set to true, use ssr for preview.